Updated product-service etc

+++ Core/Entites/OrderItem.cs
@@ -9,11 +9,9 @@ namespace Core.Entites;
 public class OrderItem
 {
     public int Id { get; set; }
-    //public int OrderId { get; set; }
-    //public Order Order { get; set; }
     public int ProductId { get; set; }
     public Product Product { get; set; }
     public int Quantity { get; set; }
-    public decimal Price { get; set; }
+    public long Price { get; set; }
 }
